Began removing unused/obsolete tests
- `ErrorMiddlewarePipe` is obsolete, and thus removed, along with its tests.
- `ApplicationMarshalLazyInteropServiceTest` is covered by the
  `MarshalMiddlewareTraitTest`.
- `ApplicationMarshalMiddlewarePipeTest` is covered by the
  `MarshalMiddlewareTraitTest`.
- Renamed `test/RouteMiddlewareTest.php` to `test/Router/IntegrationTest.php`,
  and updated it to only test integration with supported routers.
- Application now contains changes related to testing.
